清理冗余代码，抽离通用方法
