+++ src/main/java/com/ximingxing/blog/server/common/Constant.java
@@ -1,11 +1,7 @@
 package com.ximingxing.blog.server.common;
 
-import com.ximingxing.blog.server.dao.UserMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-
 public class Constant {
 
     public static Integer PAGESIZE = 15;
-
 
 }
