+++ src/main/java/com/ximingxing/blog/server/common/ServerResponse.java
@@ -2,7 +2,6 @@ package com.ximingxing.blog.server.common;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import java.io.Serializable;
