+++ src/main/java/com/ximingxing/blog/server/controller/admin/UserController.java
@@ -6,14 +6,11 @@ import com.ximingxing.blog.server.pojo.User;
 import com.ximingxing.blog.server.service.UserService;
 import com.ximingxing.blog.server.vo.UserVo;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpRequest;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/users")
