+++ src/main/java/com/ximingxing/blog/server/service/impl/RoomServiceImpl.java
@@ -7,25 +7,17 @@ import com.ximingxing.blog.server.dao.UserMapper;
 import com.ximingxing.blog.server.pojo.Room;
 import com.ximingxing.blog.server.pojo.User;
 import com.ximingxing.blog.server.service.RoomService;
+import com.ximingxing.blog.server.utils.GeneralUtils;
 import com.ximingxing.blog.server.utils.RoomUtils;
 import com.ximingxing.blog.server.utils.UserUtils;
 import com.ximingxing.blog.server.vo.RoomVo;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.poi.xssf.usermodel.XSSFRow;
-import org.apache.poi.xssf.usermodel.XSSFSheet;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.text.ParseException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -45,45 +37,21 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public ServerResponse<List<RoomVo>> uploadFile(MultipartFile file, Integer userId) {
-        String fileName;
-        fileName = file.getOriginalFilename();
-
-        String filePath;
-        String os = System.getProperty("os.name");
-        if (os.toLowerCase().startsWith("win")) {
-            filePath = "c:\\RoomsTempFiles\\";
-        } else {
-            filePath = "/Users/RoomsTempFiles/";
-        }
-
-        ExistFilePath(filePath);
-
-        File dest = new File(filePath + GetCurrentData("yyyy-MM-dd_HH-mm-ss") + fileName);
-
-        try {
-            file.transferTo(dest);
-            log.info("上传文件成功");
-        } catch (IOException e) {
-            log.error(e.toString(), e);
+        File dest = GeneralUtils.uploadFile(file);
+        if (null == dest) {
             return ServerResponse.createByError("上传失败");
         }
 
-        List<Room> list;
-        try {
-            list = excelToRoomList(dest.getPath());
-            log.info("解析文件成功");
-        } catch (ParseException e) {
-            log.info("解析文件失败");
-            e.printStackTrace();
-            return ServerResponse.createByError("解析文件失败");
+        List<Room> list = RoomUtils.analyseFileToList(dest);
+        if (list == null) {
+            return ServerResponse.createByError("解析为空");
         }
-
-        log.info("list: " + list.toString());
+        log.info("list: " + list);
 
         for (Room room : list) {
             room.setUserId(userId);
         }
-        int iCount = InsertRoomListIntoSQL(list);
+        int iCount = insertRoomListIntoSQL(list);
 
         if (0 == iCount) {
             log.info("插入数据库失败");
@@ -188,7 +156,7 @@ public class RoomServiceImpl implements RoomService {
      *
      * @param list 会议室列表
      */
-    private int InsertRoomListIntoSQL(List<Room> list) {
+    private int insertRoomListIntoSQL(List<Room> list) {
         int iCount = 0;
         for (Room r : list) {
             iCount += roomMapper.insertSelective(r);
@@ -196,81 +164,4 @@ public class RoomServiceImpl implements RoomService {
         return iCount;
     }
 
-    /**
-     * 确保路径存在
-     *
-     * @param filePath 文件夹路径
-     */
-    private void ExistFilePath(String filePath) {
-        File file = new File(filePath);
-        if (file.exists()) {
-            if (file.isDirectory()) {
-                log.info("上传目录存在");
-            } else {
-                log.info("同名文件存在, 不能创建");
-            }
-        } else {
-            log.info("目录不存在，创建目录");
-            file.mkdir();
-        }
-    }
-
-
-    /**
-     * Created By nzz
-     * 本地方法，分析Excel得出会议室信息列表
-     *
-     * @param filePath 文件路径
-     * @return List<Room>
-     */
-    private List<Room> excelToRoomList(String filePath) throws ParseException {
-
-        List<Room> list = new ArrayList<Room>();
-
-        try {
-
-            XSSFWorkbook wb = new XSSFWorkbook(new FileInputStream(filePath));
-
-            XSSFSheet sheet = wb.getSheetAt(0);
-            if (null != sheet) {
-                Room room = null;
-                for (int rowNum = 1; rowNum <= sheet.getLastRowNum(); ++rowNum) {
-                    XSSFRow row = sheet.getRow(rowNum);
-                    if (null == row) {
-                        continue;
-                    }
-                    room = new Room();
-
-                    // 获取单元格生肉值
-                    room.setRoomName(row.getCell(0).getRawValue());
-                    room.setRoomDesc(row.getCell(1).getRawValue());
-                    // 强制转化数字
-                    room.setRoomNums((int) row.getCell(2).getNumericCellValue());
-                    room.setRoomStatus(new Byte("0"));
-
-                    log.info("room: " + room.toString());
-                    list.add(room);
-                }
-            }
-            wb.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-
-        return list;
-    }
-
-
-    /**
-     * 获得当前时间
-     *
-     * @param s 时间格式
-     * @return 当前时间字符串
-     */
-    private String GetCurrentData(String s) {
-        LocalDateTime dateTime = LocalDateTime.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern(s);
-        return dateTime.format(formatter);
-    }
 }
